requirements.txt
annotated-types~=0.7.0
alembic~=1.16.2
httpx~=0.28.1
pydantic~=2.11.7
pydantic-extra-types~=2.10.5
pydantic-settings~=2.10.1
pydantic_core~=2.33.2
python-dotenv~=1.1.1
uvicorn~=0.35.0
sqlalchemy~=2.0.41
asyncpg~=0.30.0
pytest~=8.4.1
pytest-asyncio~=1.0.0
cohere~=5.16.1
pandas~=2.3.1
numpy~=2.3.2
fastapi~=0.115.14
psycopg2-binary~=2.9.10
